chat/requirements.txt
amqp==2.3.2
billiard==3.5.0.3
celery==4.1.1
colorama==0.3.9
Django==2.0.6
django-cors-headers==2.2.0
django-notifs==2.6.0
django-templated-mail==1.1.1
djangorestframework==3.8.2
djoser==1.1.5
kombu==4.2.1
pika==0.11.2
pytz==2018.4
six==1.11.0
termcolor==1.1.0
tldr==0.4.4
vine==1.1.4
